Assignment 2
Add a database called "Project1" to PostgreSQL. 
Configure the Django settings to add "PythonClub" to the app list and to make PostgreSQL with Project1 your default database. 
Edit the URL in the Project1 folder to import the URLS from the club folder. 
Add a urls.py file to the PythonClub folder. 
Migrate the database.